chore: move cookie util package
