+++ src/main/java/org/cotato/csquiz/api/auth/controller/AuthController.java
@@ -15,7 +15,7 @@ import org.cotato.csquiz.api.auth.dto.ReissueResponse;
 import org.cotato.csquiz.api.auth.dto.SendEmailRequest;
 import org.cotato.csquiz.api.member.dto.MemberEmailResponse;
 import org.cotato.csquiz.domain.auth.service.AuthService;
-import org.cotato.csquiz.domain.auth.util.CookieUtil;
+import org.cotato.csquiz.common.util.CookieUtil;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CookieValue;
